fix: 좋아요 조회를 클라이언트로 옮겨 글 페이지 SSG 복원
서버 측 getLikeBySlug의 cookies() 호출이 라우트를 동적 렌더링으로
강등시켜 generateStaticParams가 무효화된 상태였음. Reaction이
마운트 시 클라이언트에서 조회하도록 변경하고, 미사용이 된
쿠키 기반 서버 클라이언트와 @supabase/ssr 의존성 제거
