+++ package.json
@@ -23,7 +23,6 @@
     "@mdx-js/react": "^3.1.1",
     "@next/mdx": "^16.2.12",
     "@next/third-parties": "^16.2.12",
-    "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.110.9",
     "canvas-confetti": "^1.9.4",
     "clsx": "^2.1.1",
