--- src/lib/supabase/server/like.ts
@@ -1,15 +1,0 @@
-import { createClient } from '@/lib/supabase/server';
-
-export const getLikeBySlug = async (slug: string) => {
-  const supabase = await createClient();
-
-  const { data, error } = await supabase
-    .from('likes')
-    .select('count')
-    .eq('slug', slug)
-    .maybeSingle();
-
-  if (error) throw error;
-
-  return data?.count ?? 0;
-};
